requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
PyJWT==1.7.1
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.23
zipp==3.15.0
